requirements.txt
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
scikit-learn==1.2.2
flask==2.3.2
flask-cors==4.0.0
gdown==4.7.1
numpy==1.24.3
gunicorn==20.1.0